# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: "DevHub API"
  version: "0.1.0-v1.0-bootstrap"
  description: |
    DevHub v1.0 출시 직전 bootstrap. 1차 MVP scope:
    envelope 규약 + health + /me + /auth/logout + /metrics.
    도메인 endpoint 100+ (platform, repository, project, integration, DREQ 등)는
    별도 sprint — 본 spec 에 미포함.

    ## Envelope 규약
    - 성공: `{schema_version, data, request_id, trace_id}`
    - 실패: `{schema_version, data: null, request_id, trace_id, error: {code, message, details?}}`
    - `/health` 는 envelope 미사용 (raw 200).

    ## 인증
    - Bearer token (Keycloak OIDC access token, ADR-0019).
    - `/me`, `/api/v1/auth/logout` 만 `security: [{bearerAuth}]` 적용.

    ## 참고 — swagger self-link
    본 파일은 `docs/openapi.yaml` 로 직접 hand-maintain.
    Swagger UI 에서 self-link 로 표시할 경우 `/docs/openapi.yaml` 참조.
servers:
  - url: "http://localhost:8080"
    description: "local dev"
...
paths:
...
  /api/v1/admin/api-keys:
    post:
      tags:
        - Auth
      summary: "신규 API key 발급 — raw key 1회 응답"
      operationId: createApiKey
      description: |
        system_admin 일임 (routePermissionTable 정합). server-side 가 32 byte
        crypto/rand 로 raw key 생성 후 sha256 hash 만 DB 저장. 응답의 `key`
        field 는 **1회만** 노출 — frontend 가 modal 에서 clipboard copy 후
        폐기해야. 이후 GET 은 key_prefix 만 노출 (보안 invariant).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyCreateRequest"
      responses:
        "201":
          description: "발급 성공. raw key 1회 노출."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreateResponse"
        "400":
          $ref: "#/components/responses/BadRequestEnvelope"
        "401":
          $ref: "#/components/responses/UnauthorizedEnvelope"
        "403":
          $ref: "#/components/responses/ForbiddenEnvelope"
        "503":
          description: "API key store 미설정 (DEVHUB_API_KEY_STORE_PG unset)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeError"
    get:
      tags:
        - Auth
      summary: "API key 목록 조회 (key_prefix 만, raw key 미포함)"
      operationId: listApiKeys
      description: |
        모든 활성/회수/만료 키의 metadata 조회. raw key 는 절대 노출 안 됨
        (key_hash 도 미포함). system_admin 일임.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: "목록 조회 성공"
          content:
            application/json:
              schema:
                type: object
                required: [api_keys, count]
                properties:
                  api_keys:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApiKeySummary"
                  count:
                    type: integer
        "401":
          $ref: "#/components/responses/UnauthorizedEnvelope"
        "403":
          $ref: "#/components/responses/ForbiddenEnvelope"
        "503":
          $ref: "#/components/responses/ServiceUnavailableEnvelope"
  /api/v1/admin/api-keys/{api_key_id}:
    patch:
      tags:
        - Auth
      summary: "API key metadata 갱신 (expires_at / allowed_cidrs)"
      operationId: updateApiKeyMeta
      description: |
        name 은 변경 불가 (의도적). expires_at 와 allowed_cidrs 만 갱신.
        revoked key 는 update 불가 (404). system_admin 일임.
      security:
        - bearerAuth: []
      parameters:
        - name: api_key_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyUpdateRequest"
      responses:
        "200":
          description: "갱신 성공"
          content:
            application/json:
              schema:
                type: object
                required: [status, api_key_id]
                properties:
                  status:
                    type: string
                    enum: [updated]
                  api_key_id:
                    type: string
                    format: uuid
        "400":
          $ref: "#/components/responses/BadRequestEnvelope"
        "401":
          $ref: "#/components/responses/UnauthorizedEnvelope"
        "403":
          $ref: "#/components/responses/ForbiddenEnvelope"
        "404":
          $ref: "#/components/responses/NotFoundEnvelope"
        "503":
          $ref: "#/components/responses/ServiceUnavailableEnvelope"
    delete:
      tags:
        - Auth
      summary: "API key 회수 (idempotent)"
      operationId: revokeApiKey
      description: |
        ADR-0029 §6 (f) P3. 이미 회수된 key 도 audit 만 갱신 (idempotent).
        회수 후 auth middleware 가 401 reject.
      security:
        - bearerAuth: []
      parameters:
        - name: api_key_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: "회수 성공 (또는 이미 회수됨 — idempotent)"
        "401":
          $ref: "#/components/responses/UnauthorizedEnvelope"
        "403":
          $ref: "#/components/responses/ForbiddenEnvelope"
        "404":
          $ref: "#/components/responses/NotFoundEnvelope"
        "503":
          $ref: "#/components/responses/ServiceUnavailableEnvelope"
...
components:
  schemas:
...
    EnvelopeError:
      type: object
      required:
        - schema_version
        - data
        - request_id
        - trace_id
        - error
      properties:
        schema_version:
          type: string
          example: "1.0"
        data:
          type: "null"
          description: "실패 시 항상 null"
        request_id:
          type: string
          example: "req_abc123"
        trace_id:
          type: string
          example: "trace_xyz789"
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
              description: "사람 가독 에러 메시지"
              example: "invalid bearer token"
            details:
              type: string
              description: "추가 디버그 정보 (선택)"

    # -----------------------------------------------------------------------
    # UserSummary — /me data payload
    # -----------------------------------------------------------------------
...
    ErrorCode:
      type: string
      enum:
        - "auth_failed"
        - "auth_row_denied"
        - "internal_error"
        - "oidc_config_missing"
        - "oidc_network_unreachable"
        - "request_id_invalid"
        - "validation_failed"
        - "unauthenticated"
        - "not_found"
        - "unavailable"
      description: |
        에러 코드 enum (subset — 도메인 endpoint 확장 시 추가 예정).
        - auth_failed: Bearer token 검증 실패
        - auth_row_denied: RBAC row-level deny
        - internal_error: 예상치 못한 서버 오류
        - oidc_config_missing: OIDC backend config 결함
        - oidc_network_unreachable: Keycloak network/5xx
        - request_id_invalid: 잘못된 X-Request-ID 형식
        - validation_failed: 요청 payload 검증 실패
        - unauthenticated: 인증 정보 없음
        - not_found: 리소스 미존재
        - unavailable: 서비스 일시 불가

    # -----------------------------------------------------------------------
    # API Key (ADR-0029 §6 (f) P3) — multi-key 관리 schema
    # sprint: feat/work_260609-k-api-key-management
    # raw key 는 ApiKeyCreateResponse.key 에 1회만 노출 — 이후 GET 은 절대 안 됨
    # -----------------------------------------------------------------------
    ApiKeyCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
          description: "운영자 식별용 이름 (revoke/create 후 재사용 가능)"
          example: "ci-runner-prod"
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: "만료 시각 (RFC3339). null = 무기한. 미래 시각만 허용."
          example: "2027-01-01T00:00:00Z"
        allowed_cidrs:
          type: array
          items:
            type: string
          description: |
            CIDR allowlist. null/empty = all IPs 허용. auth middleware 가
            ClientIP 와 CIDR match 검증. invalid CIDR 면 400.
          example: ["10.0.0.0/8", "192.168.0.0/16"]
    ApiKeyCreateResponse:
      type: object
      required: [api_key, key, warning]
      properties:
        api_key:
          $ref: "#/components/schemas/ApiKeySummary"
        key:
          type: string
          description: |
            **raw API key — 1회만 응답.** `dhk_` prefix + base64url(32 byte) =
            47 자. frontend 가 modal 에서 clipboard copy 후 폐기. 서버는 hash
            만 보관하므로 이후 GET 으로 재조회 불가.
          example: "dhk_3J7xRt0...총47자..."
        key_expires_at:
          type: string
          format: date-time
          nullable: true
          description: "만료 시각 (request 의 expires_at 와 동일)"
        warning:
          type: string
          description: "보안 경고 문구 (한국어/영어)"
          example: "이 키는 다시 볼 수 없습니다. 안전한 곳에 즉시 저장하세요."
    ApiKeySummary:
      type: object
      required: [id, name, key_prefix, created_by, created_at, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        key_prefix:
          type: string
          description: |
            raw key 의 앞 8자 (예: `dhk_3J7x`). 운영자 식별/audit 용. raw
            key 복원에 사용 불가 (sha256 1-way).
        created_by:
          type: string
          description: "발급한 actor login"
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
        revoked_by:
          type: string
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        allowed_cidrs:
          type: array
          items:
            type: string
          nullable: true
        status:
          type: string
          enum: [active, revoked, expired]
          description: |
            `revoked_at IS NULL AND (expires_at IS NULL OR expires_at > now())`
            의 derived status. DB 컬럼이 아닌 view-time 계산.
    ApiKeyUpdateRequest:
      type: object
      properties:
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: "갱신할 만료 시각. null = clear (무기한). 미래 시각만 허용."
        allowed_cidrs:
          type: array
          items:
            type: string
          nullable: true
          description: "갱신할 CIDR allowlist. empty array = clear (all IPs)."

  # =========================================================================
  # Security Schemes
  # =========================================================================
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Keycloak-issued access token (ADR-0019)"

  # =========================================================================
  # Parameters
  # =========================================================================
...
  headers:
    XRequestID:
      description: "Correlation ID — 요청과 동일한 값 또는 서버 생성"
      schema:
        type: string
        example: "req_caller_supplied_42"
...
  responses:
    UnauthorizedEnvelope:
      description: "인증 실패 — Bearer token 없음 또는 검증 실패"
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EnvelopeError"
          example:
            schema_version: "1.0"
            data: null
            request_id: "req_unauth_001"
            trace_id: "trace_unauth_001"
            error:
              code: "unauthenticated"
              message: "invalid bearer token"
    ForbiddenEnvelope:
      description: "권한 부족 — role/permission deny"
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EnvelopeError"
          example:
            schema_version: "1.0"
            data: null
            request_id: "req_forbidden_001"
            trace_id: "trace_forbidden_001"
            error:
              code: "auth_row_denied"
              message: "insufficient role for this resource"
    BadRequestEnvelope:
      description: "잘못된 요청"
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EnvelopeError"
          example:
            schema_version: "1.0"
            data: null
            request_id: "req_bad_001"
            trace_id: "trace_bad_001"
            error:
              code: "validation_failed"
              message: "request payload validation failed"
              details: "field 'email' is required"
...
    NotFoundEnvelope:
      description: "리소스 없음 — 404"
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EnvelopeError"
          example:
            schema_version: "1.0"
            data: null
            request_id: "req_notfound_001"
            trace_id: "trace_notfound_001"
            error:
              code: "not_found"
              message: "api key not found"
    ServiceUnavailableEnvelope:
      description: "서비스 사용 불가 — 503 (api key store 미설정 등)"
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EnvelopeError"
          example:
            schema_version: "1.0"
            data: null
            request_id: "req_unavail_001"
            trace_id: "trace_unavail_001"
            error:
              code: "unavailable"
              message: "api key store is not configured"